campus-project-board/src/main/java/com/campus/projectboard/service/ArticleService.java
package com.campus.projectboard.service;

import com.campus.projectboard.dto.ArticleDto;
import com.campus.projectboard.repository.ArticleRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@RequiredArgsConstructor
@Transactional
@Service
public class ArticleService {

  private final ArticleRepository articleRepository;

  @Transactional(readOnly = true)
  public Page<ArticleDto> searchArticles(Object o, Object o1, Pageable pageable) {
    return Page.empty();
  }
}
